// frontend/src/app/players/[playerId]/types.ts
export type Player = {
  player_id: number;
  first_last: string;
  first_name: string;
  last_name: string;
  birth_date: string;
  school: string;
  country: string;
  height: string;
  weight: string;
  jersey: string;
  position: string;
  is_active: boolean;
  from_year: number;
  to_year: number;
  team_id: number;
  draft_year: string;
  draft_round: string;
  draft_number: string;
  image_url: string;
};

export type CareerStats = {
  PERSON_ID: number;
  SEASON_ID: string;
  LEAGUE_ID: string;
  TEAM_ID: number;
  TEAM_ABBREVIATION: string;
  PLAYER_AGE: number;
  GP: number;
  GS: number;
  MIN: number;
  FGM: number;
  FGA: number;
  FG_PCT: number;
  FG3M: number;
  FG3A: number;
  FG3_PCT: number;
  FTM: number;
  FTA: number;
  FT_PCT: number;
  OREB: number;
  DREB: number;
  REB: number;
  AST: number;
  STL: number;
  BLK: number;
  TOV: number;
  PF: number;
  PTS: number;
};

export type PlayerGameLog = {
  SEASON_ID: string;
  PERSON_ID: number;
  GAME_ID: string;
  GAME_DATE: string;
  MATCHUP: string;
  WL: string;
  MIN: number;
  FGM: number;
  FGA: number;
  FG_PCT: number;
  FG3M: number;
  FG3A: number;
  FG3_PCT: number;
  FTM: number;
  FTA: number;
  FT_PCT: number;
  OREB: number;
  DREB: number;
  REB: number;
  AST: number;
  STL: number;
  BLK: number;
  TOV: number;
  PF: number;
  PTS: number;
  PLUS_MINUS: number;
  VIDEO_AVAILABLE: Boolean;
};